requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.11
attrs==24.2.0
aws-lambda-powertools==2.9.0
aws-xray-sdk==2.11.0
blinker==1.8.2
boto3==1.24.88
botocore==1.27.96
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.11.0
coverage==7.6.1
deprecation==2.1.0
dnspython==2.6.1
Flask==3.0.3
functions-framework==3.8.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mock==5.1.0
packaging==24.1
pi-ina219==1.4.1
pluggy==1.5.0
py==1.11.0
pymongo==4.3.3
pytest==6.2.5
pytest-cov==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
requests==2.32.3
s3transfer==0.6.2
six==1.16.0
spidev==3.6
toml==0.10.2
typing_extensions==4.12.2
urllib3==1.26.20
watchdog==5.0.2
Werkzeug==3.0.4
wrapt==1.16.0